signess fix

+++ src/linear/blas.cpp
@@ -5,44 +5,51 @@ namespace sanity::linear::blas
 Real dot(VectorConstView x, VectorConstView y)
 {
     assert(x.size() == y.size());
-    return cblas_ddot(x.size(), &x(0), x.inc(), &y(0), y.inc());
+    return cblas_ddot((int)x.size(), &x(0), (int)x.inc(), &y(0),
+                      (int)y.inc());
 }
 
-Real nrm2(VectorConstView x) { return cblas_dnrm2(x.size(), &x(0), x.inc()); }
+Real nrm2(VectorConstView x)
+{
+    return cblas_dnrm2((int)x.size(), &x(0), (int)x.inc());
+}
 
-Real asum(VectorConstView x) { return cblas_dasum(x.size(), &x(0), x.inc()); }
+Real asum(VectorConstView x)
+{
+    return cblas_dasum((int)x.size(), &x(0), (int)x.inc());
+}
 
 int iamax(VectorConstView x)
 {
-    return cblas_idamax(x.size(), &x(0), x.inc());
+    return cblas_idamax((int)x.size(), &x(0), (int)x.inc());
 }
 
 void swap(VectorMutableView x, VectorMutableView y)
 {
     assert(x.size() == y.size());
-    cblas_dswap(x.size(), &x(0), x.inc(), &y(0), y.inc());
+    cblas_dswap((int)x.size(), &x(0), (int)x.inc(), &y(0), (int)y.inc());
 }
 
 void copy(VectorConstView x, VectorMutableView y)
 {
     assert(x.size() == y.size());
-    cblas_dcopy(x.size(), &x(0), x.inc(), &y(0), y.inc());
+    cblas_dcopy((int)x.size(), &x(0), (int)x.inc(), &y(0), (int)y.inc());
 }
 
 void axpy(Real a, VectorConstView x, VectorMutableView y)
 {
     assert(x.size() == y.size());
-    cblas_daxpy(x.size(), a, &x(0), x.inc(), &y(0), y.inc());
+    cblas_daxpy((int)x.size(), a, &x(0), (int)x.inc(), &y(0), (int)y.inc());
 }
 
 void scal(Real a, VectorMutableView x)
 {
-    cblas_dscal(x.size(), a, &x(0), x.inc());
+    cblas_dscal((int)x.size(), a, &x(0), (int)x.inc());
 }
 
 void scal(Complex a, CVectorMutableView x)
 {
-    cblas_zscal(x.size(), &a, &x(0), x.inc());
+    cblas_zscal((int)x.size(), &a, &x(0), (int)x.inc());
 }
 
 static CBLAS_ORDER order = CblasRowMajor;
@@ -72,8 +79,9 @@ void gemv(Real a, MatrixConstView A, Oper op, VectorConstView x, Real b,
         assert(A.nrow() == x.size());
         assert(A.ncol() == y.size());
     }
-    cblas_dgemv(order, blasOper(op), A.nrow(), A.ncol(), a, &A(0, 0),
-                A.ldim(), &x(0), x.inc(), b, &y(0), y.inc());
+    cblas_dgemv(order, blasOper(op), (int)A.nrow(), (int)A.ncol(), a,
+                &A(0, 0), (int)A.ldim(), &x(0), (int)x.inc(), b, &y(0),
+                (int)y.inc());
 }
 
 void gemv(Complex a, CMatrixConstView A, Oper op, CVectorConstView x,
@@ -89,14 +97,15 @@ void gemv(Complex a, CMatrixConstView A, Oper op, CVectorConstView x,
         assert(A.nrow() == x.size());
         assert(A.ncol() == y.size());
     }
-    cblas_zgemv(order, blasOper(op), A.nrow(), A.ncol(), &a, &A(0, 0),
-                A.ldim(), &x(0), x.inc(), &b, &y(0), y.inc());
+    cblas_zgemv(order, blasOper(op), (int)A.nrow(), (int)A.ncol(), &a,
+                &A(0, 0), (int)A.ldim(), &x(0), (int)x.inc(), &b, &y(0),
+                (int)y.inc());
 }
 
 void gemm(Real a, MatrixConstView A, Oper opA, MatrixConstView B, Oper opB,
           Real b, MatrixMutableView C)
 {
-    int m, n, k;
+    uint m, n, k;
     if (opA == NoTranspose)
     {
         m = A.nrow();
@@ -119,8 +128,9 @@ void gemm(Real a, MatrixConstView A, Oper opA, MatrixConstView B, Oper opB,
     }
     assert(m == C.nrow());
     assert(n == C.ncol());
-    cblas_dgemm(order, blasOper(opA), blasOper(opB), m, n, k, a, &A(0, 0),
-                A.ldim(), &B(0, 0), B.ldim(), b, &C(0, 0), C.ldim());
+    cblas_dgemm(order, blasOper(opA), blasOper(opB), (int)m, (int)n, (int)k,
+                a, &A(0, 0), (int)A.ldim(), &B(0, 0), (int)B.ldim(), b,
+                &C(0, 0), (int)C.ldim());
 }
 
 }  // namespace sanity::linear::blas
