+++ src/powerflow/solve.hpp
@@ -9,20 +9,20 @@ namespace sanity::powerflow
 {
 class BusMap
 {
-    std::vector<int> _grid2mat;
-    std::vector<int> _mat2grid;
+    std::vector<uint> _grid2mat;
+    std::vector<uint> _mat2grid;
 
 public:
-    BusMap(std::vector<int> grid2mat, std::vector<int> mat2grid)
+    BusMap(std::vector<uint> grid2mat, std::vector<uint> mat2grid)
         : _grid2mat(std::move(grid2mat)), _mat2grid(std::move(mat2grid))
     {
     }
 
-    int getGridIdx(int mat_idx) const { return _mat2grid[(uint)mat_idx]; }
-    int getMatrixIdx(int grid_idx) const { return _grid2mat[(uint)grid_idx]; }
+    uint getGridIdx(uint mat_idx) const { return _mat2grid[mat_idx]; }
+    uint getMatrixIdx(uint grid_idx) const { return _grid2mat[grid_idx]; }
 };
 
-BusMap remapBus(const PowerGrid& grid, int slackBus);
+BusMap remapBus(const PowerGrid& grid, uint slackBus);
 linear::CMatrix admittanceMatrix(const PowerGrid& grid,
                                  const BusMap& bus_map);
 void fullJacobian(linear::VectorConstView VAngle,
@@ -36,13 +36,13 @@ struct BusState
     Complex voltage;
 };
 
-std::vector<BusState> flatStart(const PowerGrid& grid, int slackBusIdx);
+std::vector<BusState> flatStart(const PowerGrid& grid, uint slackBusIdx);
 
 // sol is the init guess, which will be overwritten with the final solution.
 IterationResult solveNewton(
-    const PowerGrid& grid, int slackBusIdx, std::vector<BusState>& sol,
+    const PowerGrid& grid, uint slackBusIdx, std::vector<BusState>& sol,
     const std::function<void(linear::MatrixMutableView A,
                              linear::VectorMutableView xb)>& linear_solver,
-    int max_iter, Real tol);
+    uint max_iter, Real tol);
 
 }  // namespace sanity::powerflow
